+++ src/Zicht/Bundle/KeyValueBundle/Command/KeyValueMigrateJsonSchemaKeysCommand.php
@@ -8,13 +8,12 @@ namespace Zicht\Bundle\KeyValueBundle\Command;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Helper\Table;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
 use Zicht\Bundle\KeyValueBundle\KeyValueStorage\KeyValueStorageManager;
 use Zicht\Bundle\KeyValueBundle\KeyValueStorage\PredefinedJsonSchemaKey;
-use Swaggest\JsonSchema\Exception\ObjectException;
 
 class KeyValueMigrateJsonSchemaKeysCommand extends ContainerAwareCommand
 {
@@ -26,6 +25,7 @@ class KeyValueMigrateJsonSchemaKeysCommand extends ContainerAwareCommand
         $this
             ->setName('zicht:key-value:migrate-json-schema-keys')
             ->setDescription('Validate all json schema keys and migrate when possible')
+            ->addOption('replace-invalid', null, InputOption::VALUE_NONE, 'Replace stored value with default value when unable to migrate')
             ->addOption('force', null, InputOption::VALUE_NONE, 'Force migration in the database to occur');
     }
 
@@ -34,6 +34,8 @@ class KeyValueMigrateJsonSchemaKeysCommand extends ContainerAwareCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $io = new SymfonyStyle($input, $output);
+
         $table = new Table($output);
         $table->setHeaders(['Key', 'Default', 'Storage', 'Migrate']);
 
@@ -47,16 +49,28 @@ class KeyValueMigrateJsonSchemaKeysCommand extends ContainerAwareCommand
             if ($predefinedKey instanceof PredefinedJsonSchemaKey) {
                 $key = $predefinedKey->getKey();
                 $defaultValue = $predefinedKey->getValue();
-                $defaultValueIsValid = $predefinedKey->isValid($defaultValue);
+                if (!$defaultValueIsValid = $predefinedKey->isValid($defaultValue, $message)) {
+                    $io->writeln(sprintf('%s - validation error on default value', $key));
+                    $io->writeln($message);
+                    $io->error(json_encode($defaultValue, JSON_PRETTY_PRINT));
+                }
                 $defaultState = $defaultValueIsValid ? 'ok' : '<error>invalid</error>';
                 $storageValue = $storageManager->getValue($key);
-                $storageValueIsValid = $predefinedKey->isValid($storageValue);
+                if (!$storageValueIsValid = $predefinedKey->isValid($storageValue, $message)) {
+                    $io->writeln(sprintf('%s - validation error on storage value', $key));
+                    $io->writeln($message);
+                    $io->error(json_encode($storageValue, JSON_PRETTY_PRINT));
+                }
                 $storageState = $storageValueIsValid ? 'ok' : '<error>invalid</error>';
                 $migrateState = $storageValueIsValid ? '' : '<error>unable to migrate</error>';
 
                 if ($defaultValueIsValid && !$storageValueIsValid) {
-                    $migrateValue = array_replace_recursive($defaultValue, $storageValue);
-                    $migrateValueIsValid = $predefinedKey->isValid($migrateValue);
+                    $migrateValue = $predefinedKey->migrate($storageValue);
+                    if (!$migrateValueIsValid = $predefinedKey->isValid($migrateValue, $message)) {
+                        $io->writeln(sprintf('%s - validation error on migrated value', $key));
+                        $io->writeln($message);
+                        $io->error(json_encode($migrateValue, JSON_PRETTY_PRINT));
+                    }
 
                     if ($migrateValueIsValid) {
                         if ($input->getOption('force')) {
@@ -66,6 +80,17 @@ class KeyValueMigrateJsonSchemaKeysCommand extends ContainerAwareCommand
                             $migrateState = 'migrated';
                         } else {
                             $migrateState = 'use --force to migrate';
+                        }
+                    } else {
+                        if ($input->getOption('replace-invalid')) {
+                            if ($input->getOption('force')) {
+                                $storageManager->saveValue($key, $defaultValue);
+                                $doctrine->getEntityManager()->flush();
+                                $storageState = strip_tags($storageState);
+                                $migrateState = 'replaced with default';
+                            } else {
+                                $migrateState = 'use --force to replace with default';
+                            }
                         }
                     }
                 }